Prune dead code, clear lint backlogs and remove defensive audit machinery

+++ robocon-mid360-autonomy-stack/src/robocon_mid360_simulation/scripts/depth_visualizer.py
@@ -54,7 +54,7 @@ class DepthVisualizer(Node):
                 normalized = (self.max_depth - min(self.max_depth, value)) / (
                     self.max_depth - self.min_depth
                 )
-                output[row * width + column] = int(round(255.0 * normalized))
+                output[row * width + column] = round(255.0 * normalized)
 
         visualized = Image()
         visualized.header = message.header
